# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Dou Dizhu Backend API
  version: 1.0.0
  description: RESTful API for Dou Dizhu multiplayer game backend
  contact:
    name: Backend Team
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://api.doudizhu.example.com
    description: Production server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: INVALID_INPUT
            message:
              type: string
              example: Invalid request data
      required:
        - success
        - error
...
    Room:
      type: object
      properties:
        roomId:
          type: string
          format: uuid
        name:
          type: string
        hostId:
          type: string
          format: uuid
        players:
          type: array
          items:
            $ref: '#/components/schemas/RoomPlayer'
        maxPlayers:
          type: integer
          example: 4
        isPrivate:
          type: boolean
        type:
          type: string
          enum: [PVP, PVE]
          default: PVP
        difficulty:
          type: string
          enum: [EASY, MEDIUM, HARD]
          default: MEDIUM
        botCount:
          type: integer
          default: 0
        status:
          type: string
          enum: [waiting, playing, finished]
        createdAt:
          type: string
          format: date-time
    RoomPlayer:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        username:
          type: string
        isReady:
          type: boolean
...
paths:
...
  /rooms:
    get:
      summary: List available game rooms
      tags:
        - Game Rooms
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [waiting, playing, finished]
          description: Filter by room status
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Room list retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      rooms:
                        type: array
                        items:
                          $ref: '#/components/schemas/Room'
                      pagination:
                        type: object
                        properties:
                          page:
                            type: integer
                          limit:
                            type: integer
                          total:
                            type: integer
                          totalPages:
                            type: integer

    post:
      summary: Create a new game room
      tags:
        - Game Rooms
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                maxPlayers:
                  type: integer
                  default: 4
                isPrivate:
                  type: boolean
                  default: false
                type:
                  type: string
                  enum: [PVP, PVE]
                  default: PVP
                difficulty:
                  type: string
                  enum: [EASY, MEDIUM, HARD]
                  default: MEDIUM
                botCount:
                  type: integer
                  default: 0
                password:
                  type: string
                  description: Required for private rooms
            example:
              name: My Game Room
              maxPlayers: 4
              isPrivate: false
      responses:
        '201':
          description: Room created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Room'
  /rooms/{roomId}:
    get:
      summary: Get room details
      tags:
        - Game Rooms
      security:
        - BearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Room details retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Room'
        '404':
          description: Room not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete room (host only)
      tags:
        - Game Rooms
      security:
        - BearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Room deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '403':
          description: Not room host
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /rooms/{roomId}/join:
    post:
      summary: Join an existing room
      tags:
        - Game Rooms
      security:
        - BearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                  description: Required for private rooms
      responses:
        '200':
          description: Joined room successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      roomId:
                        type: string
                        format: uuid
                      playerId:
                        type: string
                        format: uuid
        '403':
          description: Room is full or wrong password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Room not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /rooms/{roomId}/ai:
    post:
      summary: Add an AI bot to the room
      tags:
        - Game Rooms
      security:
        - BearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Bot added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/RoomPlayer'
        '400':
          description: Room is full
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Room not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /rooms/{roomId}/fill-bots:
    post:
      summary: Fill all empty seats with AI bots
      description: Adds AI bots to all empty seats in the room in a single request. This is the recommended endpoint for PVE games to improve UX by eliminating multiple clicks.
      tags:
        - Game Rooms
      security:
        - BearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Bots added successfully (or room already full)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Room is already full
                    description: Only present when room is already full
                  data:
                    type: object
                    properties:
                      botsAdded:
                        type: integer
                        example: 3
                        description: Number of bots that were added
                      bots:
                        type: array
                        items:
                          $ref: '#/components/schemas/RoomPlayer'
                        description: Array of bot players that were added
        '404':
          description: Room not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /rooms/{roomId}/leave:
    post:
      summary: Leave current room
      tags:
        - Game Rooms
      security:
        - BearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Left room successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
...